fix: Prevent unintended re-renders in the `TranslationContextProvider`
In React, raw/unmemoized objects can be a cause of unintended re-renders.
With the previous code, the entire application that lives inside
`TranslationContextProvider` would re-render whenever the component
itself re-renders (because the `value` object would get re-defined on
each render).

Now that we're using `useMemo`, this problem won't happen anymore.
Unfortunately, due to the Rules of Hooks, this memoization has
to be done even if nothing is being rendered to the DOM.

For additional context, see:
https://github.com/jsx-eslint/eslint-plugin-react/blob/master/docs/rules/jsx-no-constructed-context-values.md